# Faster JSON serialization