--- a/Elements.Benchmarks/Trace.cs
+++ b/Elements.Benchmarks/Trace.cs
@@ -1,3 +1,4 @@
+using System.IO;
 using System.Linq;
 using BenchmarkDotNet.Attributes;
 using BenchmarkDotNet.Diagnosers;
@@ -13,10 +14,11 @@ namespace Elements.Benchmarks
         [Benchmark(Description = "Create all HSS beams and serialize to JSON.")]
         public void TraceModelCreation()
         {
+            Validators.Validator.DisableValidationOnConstruction = true;
             var factory = new HSSPipeProfileFactory();
             var hssProfiles = factory.AllProfiles().ToList();
             var model = ElementCreation.DrawAllBeams(hssProfiles);
-            model.ToJson();
+            model.ToJson(gatherSubElements: false);
         }
     }