--- a/Elements.Benchmarks/Trace.cs
+++ b/Elements.Benchmarks/Trace.cs
@@ -1,4 +1,3 @@
-using System.IO;
 using System.Linq;
 using BenchmarkDotNet.Attributes;
 using BenchmarkDotNet.Diagnosers;