BACKGROUND:
- During updates to Hypar's structure function, I noticed that JSON serialization was 10x slower than the creation of all of the elements. After some investigation I found that our injection of discriminators into `JObject` types created using a Json.net API is more than half the time spent writing to JSON. This is currently unavoidable without a large refactoring of how we handle discriminators. There were a number of small optimizations that we could make however which removed ~20ms from our JSON serialization benchmark (even before I converted that benchmark to disable validation and not use recursive element addition). 

DESCRIPTION:
- See comments inline about specific improvements.

TESTING:
- See the updated benchmark below.
  
FUTURE WORK:
- Is there any future work needed or anticipated?  Does this PR create any obvious technical debt?

REQUIRED:
- [x] All changes are up to date in `CHANGELOG.md`. **n/a** 

Before:
|                                        Method |     Mean |   Error |  StdDev |
|---------------------------------------------- |---------:|--------:|--------:|
| 'Create all HSS beams and serialize to JSON.' | 124.4 ms | 2.47 ms | 3.85 ms |

After:
|                                        Method |     Mean |   Error |  StdDev |
|---------------------------------------------- |---------:|--------:|--------:|
| 'Create all HSS beams and serialize to JSON.' | **104.7 ms** | 2.07 ms | 3.68 ms |

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/hypar-io/elements/749)
<!-- Reviewable:end -->